Added gitignore and new view function
